fix: Wrap useSearchParams in Suspense boundary on Community page.
